[Add/#461] ORBMapHelper, PlaceInfoToolipHelper 추가 및 지도, 툴팁 애니메이션 역할 분리
ORBMapHelper와 PlaceInfoToolipHelper 타입을 정의하고, 해당 타입에 각각 지도 이동, 툴팁 애니메이션 역할을 분담.
